Re-implemented Unity features.

+++ LiveSplit.ASLHelper/MainHelper/Helper/Memory/Scan/Main_ScanAll.cs
@@ -3,39 +3,39 @@
 public partial class Main
 {
     #region String Pattern
-    public IEnumerable<IntPtr> ScanAll(params string[] pattern)
+    public IEnumerable<nint> ScanAll(params string[] pattern)
     {
         var module = Game?.MainModuleWow64Safe();
         return ScanAll(module, 0, pattern);
     }
 
-    public IEnumerable<IntPtr> ScanAll(int offset, params string[] pattern)
+    public IEnumerable<nint> ScanAll(int offset, params string[] pattern)
     {
         var module = Game?.MainModuleWow64Safe();
         return ScanAll(module, offset, pattern);
     }
 
-    public IEnumerable<IntPtr> ScanAll(string module, params string[] pattern)
+    public IEnumerable<nint> ScanAll(string module, params string[] pattern)
     {
         return ScanAll(GetModule(module), 0, pattern);
     }
 
-    public IEnumerable<IntPtr> ScanAll(string module, int offset, params string[] pattern)
+    public IEnumerable<nint> ScanAll(string module, int offset, params string[] pattern)
     {
         return ScanAll(GetModule(module), offset, pattern);
     }
 
-    public IEnumerable<IntPtr> ScanAll(ProcessModuleWow64Safe module, params string[] pattern)
+    public IEnumerable<nint> ScanAll(ProcessModuleWow64Safe module, params string[] pattern)
     {
         return ScanAll(module, 0, pattern);
     }
 
-    public IEnumerable<IntPtr> ScanAll(ProcessModuleWow64Safe module, int offset, params string[] pattern)
+    public IEnumerable<nint> ScanAll(ProcessModuleWow64Safe module, int offset, params string[] pattern)
     {
         if (module == null)
         {
             Debug.Warn("[Scan] Module could not be found!");
-            return Enumerable.Empty<IntPtr>();
+            return Enumerable.Empty<nint>();
         }
 
         var start = module.BaseAddress;
@@ -44,33 +44,33 @@ public partial class Main
         return ScanAll(start, size, offset, pattern);
     }
 
-    public IEnumerable<IntPtr> ScanAll(IntPtr startAddress, IntPtr endAddress, params string[] pattern)
+    public IEnumerable<nint> ScanAll(nint startAddress, nint endAddress, params string[] pattern)
     {
         var start = startAddress;
-        var size = (int)((long)endAddress - (long)startAddress);
+        var size = (int)(endAddress - (long)startAddress);
 
         return ScanAll(start, size, 0, pattern);
     }
 
-    public IEnumerable<IntPtr> ScanAll(IntPtr startAddress, IntPtr endAddress, int offset, params string[] pattern)
+    public IEnumerable<nint> ScanAll(nint startAddress, nint endAddress, int offset, params string[] pattern)
     {
         var start = startAddress;
-        var size = (int)((long)endAddress - (long)startAddress);
+        var size = (int)(endAddress - (long)startAddress);
 
         return ScanAll(start, size, offset, pattern);
     }
 
-    public IEnumerable<IntPtr> ScanAll(IntPtr startAddress, int size, params string[] pattern)
+    public IEnumerable<nint> ScanAll(nint startAddress, int size, params string[] pattern)
     {
         return ScanAll(startAddress, size, 0, pattern);
     }
 
-    public IEnumerable<IntPtr> ScanAll(IntPtr startAddress, int size, int offset, params string[] pattern)
+    public IEnumerable<nint> ScanAll(nint startAddress, int size, int offset, params string[] pattern)
     {
         if (Game == null)
         {
             Debug.Warn("[Scan] Game process was null!");
-            return Enumerable.Empty<IntPtr>();
+            return Enumerable.Empty<nint>();
         }
 
         Debug.Log(startAddress.ToString("X"));
@@ -85,23 +85,23 @@ public partial class Main
     #endregion
 
     #region SigScanTarget
-    public IEnumerable<IntPtr> ScanAll(SigScanTarget target)
+    public IEnumerable<nint> ScanAll(SigScanTarget target)
     {
         var module = Game?.MainModuleWow64Safe();
         return ScanAll(module, target);
     }
 
-    public IEnumerable<IntPtr> ScanAll(string module, SigScanTarget target)
+    public IEnumerable<nint> ScanAll(string module, SigScanTarget target)
     {
         return ScanAll(GetModule(module), target);
     }
 
-    public IEnumerable<IntPtr> ScanAll(ProcessModuleWow64Safe module, SigScanTarget target)
+    public IEnumerable<nint> ScanAll(ProcessModuleWow64Safe module, SigScanTarget target)
     {
         if (module == null)
         {
             Debug.Warn("[Scan] Module could not be found!");
-            return Enumerable.Empty<IntPtr>();
+            return Enumerable.Empty<nint>();
         }
 
         var start = module.BaseAddress;
@@ -110,20 +110,20 @@ public partial class Main
         return ScanAll(start, size, target);
     }
 
-    public IEnumerable<IntPtr> ScanAll(IntPtr startAddress, IntPtr endAddress, SigScanTarget target)
+    public IEnumerable<nint> ScanAll(nint startAddress, nint endAddress, SigScanTarget target)
     {
         var start = startAddress;
-        var size = (int)((long)endAddress - (long)startAddress);
+        var size = (int)(endAddress - (long)startAddress);
 
         return ScanAll(start, size, target);
     }
 
-    public IEnumerable<IntPtr> ScanAll(IntPtr startAddress, int size, SigScanTarget target)
+    public IEnumerable<nint> ScanAll(nint startAddress, int size, SigScanTarget target)
     {
         if (Game == null)
         {
             Debug.Warn("[Scan] Game process was null!");
-            return Enumerable.Empty<IntPtr>();
+            return Enumerable.Empty<nint>();
         }
 
         var scanner = new SignatureScanner(Game, startAddress, size);
